Use the non_exhaustive feature in CertificateParams
This attribute replaces the _hidden field of the
struct, serving the same purpose.

For enums, we don't use the non_exhaustive feature
yet as we'd like to have exhaustivness checks
at least for our internal code.
For more, see this comment by @dtolnay [1]

It stabilized in Rust 1.40.0 which got released today.
The MSRV policy of this crate is to follow whatever
the currently latest released Rust version is.

[1]: https://github.com/rust-lang/rust/issues/44109#issuecomment-521781237