Don't write a subjectAltName if no subject alt name was specified
RFC 5280 states:

    If the subjectAltName extension is present, the sequence MUST contain
    at least one entry.

Thus, if the sequence would be empty, we'd have to omit the entire
extension. Same goes for the definition of the SubjectAltName extension:

   SubjectAltName ::= GeneralNames

   GeneralNames ::= SEQUENCE SIZE (1..MAX) OF GeneralName

Note the 1. As for the entire block of extensions, we shouldn't write
them unless there is at least one extension to write. See this
definition (Also from RFC 5820):

   Extensions  ::=  SEQUENCE SIZE (1..MAX) OF Extension

The CSR stuff I'm not sure about but RFC 2985 imports the Extensions
type from ISO/IEC 9594-8:1997, but if it's the same as in RFC 5280,
then it has to have at least 1 item as well.

Fixes #29.