Generalize extension serialization and fix SubjectKeyIdentifier format
RFC 5280 states the following X.509v3 extension format:
   Extension  ::=  SEQUENCE  {
        extnID      OBJECT IDENTIFIER,
        critical    BOOLEAN DEFAULT FALSE,
        extnValue   OCTET STRING
                    -- contains the DER encoding of an ASN.1 value
                    -- corresponding to the extension type identified
                    -- by extnID
        }

For the SubjectKeyIdentifier extension, this format was not adhered,
the digest was written as an OCTET STRING directly,
instead of its DER-encoding.
This caused issues with a variety of X.509 implementations,
such as WolfSSL or the one present in Go,
which did not accept generated CA certificates.