Use RFC 7093 method 1 to calculate subject key identifiers
It does not make a lot of sense to make key identifier generation dependent on the used signature algorithm.
Ideally, key identifiers are short and uniform for lookup efficiency.
RFC 5280 originally suggests using SHA-1 for key identifiers, but since its deprecation, this is no longer a viable option.
RFC 7093 instead suggests using SHA-2 and truncating the result to 160 bits.
This also ensures compatibility with non-standard-conformant X.509 clients, which may expect such a length.