The permitted and excluded subtrees have to be implicitly tagged
The ASN.1 definition of NameConstraints is found in the implicitly
tagged section of RFC 5820 here [0]:

NameConstraints ::= SEQUENCE {
     permittedSubtrees       [0]     GeneralSubtrees OPTIONAL,
     excludedSubtrees        [1]     GeneralSubtrees OPTIONAL }

GeneralSubtrees ::= SEQUENCE SIZE (1..MAX) OF GeneralSubtree

GeneralSubtree ::= SEQUENCE {
     base                    GeneralName,
     minimum         [0]     BaseDistance DEFAULT 0,
     maximum         [1]     BaseDistance OPTIONAL }

At the start of the section, there is this definition,
setting the default tag type to implicit:

DEFINITIONS IMPLICIT TAGS ::=

I've been wondering why openssl didn't work while webpki did work.
For hours I was blaming openssl to be wrong and even prepared a
bug report for them. But thankfully I researched more and
found out that the tags are implicit in this instance.

Turns out that webpki assumed them to be explicit as well,
so I filed a bug report for them [1]. This makes two functional
limitations/bugs in the name constraints code of webpki that
we've found.

Also found a nice article by lets encrypt on explicit vs implicit
tags [2].

[0]: https://tools.ietf.org/html/rfc5280#page-129
[1]: https://github.com/briansmith/webpki/issues/135
[2]: https://letsencrypt.org/de/docs/a-warm-welcome-to-asn1-and-der/