Use public key value as a default serial number (#61)
Firefox will complain if the same serial number (defaults to 42) is used with different certificate. This happens if you regenerate the certificate. This patch uses the hash of the public key (first 8 bytes) as the value for the serial number. This way every certificate will get a unique serial number.