Make sure that all CertificateParams can be serialized in a CSR (#107)
In write_request, we now serialize custom extensions as extension
requests.  We also check that all fields are either empty or that we can
serialize them.

We do ignore a few fields: not_before, not_after, and
key_identifier_method.
NotBefore and NotAfter cannot be specified in a CSR, a key identifier
might be but would require a use case and API changes.

Fixes https://github.com/est31/rcgen/issues/64.