tests: update botan tests to include CRL support.
This commit updates the `botan-rs` test suite to include support for
parsing a CRL generated by `rcgen`, as well as checking a revoked
certificate is present in the CRL contents.

The `botan-rs` lib doesn't yet support using CRLs when validating
a certificate chain, or verifying the signature over a CRL, pending
updates to the underlying C++ `botan` lib.