project: codecov comments only for changes.
This commit adds a `codecov.yml` configuration file that instructs
Codecov.io to only add a comment to pull requests when coverage changes.

See the codecov docs[0] for more information.

If this is still too noisy for our preferences we can adjust the config
to `comment: false` in a follow-up.

[0]: https://docs.codecov.com/docs/pull-request-comments