ci: add vs-shell setup task to configure Windows env
The upstream `botan-rs` crate uses this same action in its Windows based
CI, successfully building `botan-sys` on that platform.

Without using this action, the build process fails with the error:
```
ERROR: Default compiler is msvc but could not find 'cl'; use --cc to set
```

It's possible that with effort another workaround could be used, but
it seems prudent to match what the upstream repository does and its
better to have a working setup we can iterate on than one that skips
Botan coverage entirely.