ci: pull out separate rust doc job
Previously `cargo doc` was being invoked in the per-platform CI jobs.
This feels unnecessary since no documentation items are gated
per-platform and we can rely on the upstream `cargo doc` tool owners to
test for general doc generation platform compatibility.

In this commit the `cargo doc` invocation is moved to a dedicated job.
This also allows shifting the `RUSTDOCFLAGS` env var from a global env
var to one scoped to just the job that requires it.

We continue to test doc generation using the stable, beta, nightly and
MSRV rust versions.