lib: enable `doc_auto_cfg` for docs.rs
The upstream `docs.rs` website uses the nightly release of the Rust
compiler when building project docs. That means we can opt in to the
`doc_auto_cfg` feature to have documentation items automatically mention
the required crate features where appropriate.

This has the advantages of:

* Using a nice styled UI element for the requirements.
* Keeping the feature requirements in-sync with the docs automatically.

Since this feature requires the nightly toolchain we customize the CI
environment so that it is enabled only for the nightly channel.

[0]: https://docs.rs/about/builds