ci: add cargo-check-external-types, config
This commit adds configuration and a CI task for checking that no
types from dependencies are accidentally leaked through the Rcgen public
API unintentionally.

The previous commits in this branch fixed the `*ring*` type leaks, so
our configuration only has two white-listed types as of this branch:

1. `time::offset_date_time::OffsetDateTime`

It's unclear whether usage of that type should be adjusted, so for now
we explicitly allow-list it in the cargo-check-external-types config. We
can deal with this type (or not) in the future.

2. `zeroize::Zeroize`

We could probably avoid leaking this type by implementing `Drop` and
calling `zeroize` on fields directly from the `drop` impl. In the
meantime we add this type to the allow list.