error: avoid leaking pem::PemError into api
Having `From<pem::PemError>` defined on the public `Error`
type means that the `pem` type leaks into rcgen's public API,
complicating semver incompatible updates.

This commit updates the sites that previously used this trait to
use the crate internal `ExternalError` extension trait to map the
`PemError` err to the generic rcgen `Error::PemError` err.

Additionally, the `rcgen::Error::PemError` variant is changed to hold
a `String` with the `pem::PemError` error string instead of the type
itself. This allows the `From` impl on `Error` to be removed, fixing the
type leak.