ext: implement subject alternative name.
This commit lifts the subject alternative name extension into the `ext`
module.

It additionally ensures we never write an empty SAN extension, if the
`CertificateParams` contain an empty vec of SAN names.

For the time being SAN extensions are always written as non-criticial,
but the required plumbing to handle the RFC5280 guidance on SAN ext
criticality is added for follow-up adjustment.