lib: have AKI write fn take AKI value as arg
Rather than have the `write_x509_authority_key_identifier` fn take
a `Certificate` as input in order to calculate the AKI value, have the
caller find the AKI and pass it to the write fn for writing. This
simplifies generating a certificate with the issuer's `KeyPair`, and
not a `Certificate`.