lib: return CertifiedKey from generate
Rather than store `KeyPair` in `Certificate`, take `KeyPair` as
a parameter when required for signing and return `KeyPair` from the
generate methods as a new `CertifiedKey` type. This will allow unifying
the API around the idea that a `Certificate` is public material.

Tests and the cert gen utility are updated accordingly.