lib: add clarification for CSR serialization fns
This commit updates the rustdoc comments on
`Certificate::serialize_request_der` and
`Certificate::serialize_request_pem`, emphasizing that they generate
and sign CSRs for each invocation. The purpose of the `subject_key`
argument is emphasized.

Along the way, fix an inaccurate comment in
`CertificateParams::write_request` - we're writing the _subject_
distinguished name, not the issuer. The `Name` field in the syntax in
RFC 2986 is `subject`.