Support compiling without cryptography primitives (#208)
This makes all of the crypto libraries optional. My use case requires
simply generating the binary ASN.1 structures and all cryptography is
handled externally with HSMs.

This change removes unnecessary code and build complexity (for example,
I'm targeting WASM, and I need to take extra steps to build the
dependencies I don't need).

Note to the maintainer: this is a clean (and ideally better) rework of a
previous PR, #207. Because I deleted the older and messier branch, the
previous PR couldn't be updated, so I closed it.