cert: use key_identifier_method of issuer for AKI
Previously when issuing a certificate with an authority key identifier
(AKI) extension that's signed by an issuer certificate we had a small
bug where we used the to-be-issued certificate's param's `key_identifier_method`
to derive the key identifier of the issuing certificate to use for the
issued certificate's AKI.

Instead we should be using the issuer certificate's param's
`key_identifier_method`, taking care to mind the pre-specified variant.

We missed this with our unit testing of the pre-specified key identifier
method because we only issued a self-signed test certificate, never
issuing a certificate signed by the CA that has the customization.

This commit fixes the bug and extends test coverage to prevent
further regression.